CRUD oprations and statistics
Statistics for sellers were add in project and are working fine.
The CRUD operations for sellers also are working fine, but it's possible to optimize the methods of statistics and list of all sellers in only only one, because the code are very similar, the only thing different is the sql query.